requirements.txt
# Core
python-dotenv==1.0.0
asyncio==3.4.3

# Database (usando psycopg3 - compatível com Python 3.13)
psycopg[binary]==3.1.18
sqlalchemy==2.0.23

# Data & Analysis
pandas==2.1.4
numpy==1.26.2
pandas-ta==0.3.14b0

# WebSocket
websockets==12.0
aiohttp==3.9.1

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.0

# ML & Backtest (compatível com Python 3.13)
xgboost>=2.0.0
lightgbm>=4.1.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Analysis
jupyter>=1.0.0
seaborn>=0.13.0
matplotlib>=3.8.0
